pyproject.toml
[project]
name = "brondol-app"
version = "0.1.0"
requires-python = ">=3.10"

dependencies = [
    "numpy<2",
    "pyserial>=3.5",
    "opencv-python>=4.11.0.86",
    "ultralytics>=8.3.229",
    "torch==2.0.1; sys_platform == 'darwin' and platform_machine == 'x86_64'",
    "torch==2.9.1; sys_platform == 'darwin' and platform_machine == 'arm64'",
    "torch==2.2.2; sys_platform == 'linux' and platform_machine == 'aarch64'",
]

[tool.uv]
required-environments = [
    "sys_platform == 'darwin' and platform_machine == 'arm64'",
    "sys_platform == 'darwin' and platform_machine == 'x86_64'",
]